+++ NEWS.md
@@ -1,6 +1,9 @@
 # tsibble 0.1.2.9999
 
+## API changes
+
 * The windowed functions, including `slide()`, `tile()`, `stretch()`, are no longer defined as S3 methods. Several new variants have been introduced for the purpose of type stability, like `slide_lst()` (a list), `slide_dfr()` (a row-binding data frame), `slide_dfc()` (a column-binding data frame).
+* The `index` variable must sit in the first name-value pair in `tsummarise()` instead of any position in the call.
 
 ## Bug fixes